+++ src/ServiceFabric/ServiceFabric/ChangeLog.md
@@ -19,7 +19,7 @@
 -->
 ## Upcoming Release
 * Added parameters `-EnableAutoOsUpgrde` and `-AllowRdpAccess` to `New-AzServiceFabricManagedCluster` and `Set-AzServiceFabricManagedCluster`.
-* Added parameters `-ZoneBalance`, `-AllowOverProvisioning`, and `-Zones` to `New-AzServiceFabricManagedNodeType` and `Set-AzServiceFabricManagedNodeType`.
+* Added parameters `-ZoneBalance`, `-AllowOverProvisioning`, and `-Zone` to `New-AzServiceFabricManagedNodeType` and `Set-AzServiceFabricManagedNodeType`.
 * Added preannoucement for removing Reimage parameters from `Set-AzServiceFabricManagedNodeType`.
 * Added new cmdlets for managed node types:
     - `Invoke-AzServiceFabricDeallocateManagedNodeType`